+++ utils/src/main/kotlin/org/archivekeep/utils/collections/collections.kt
@@ -1,0 +1,7 @@
+package org.archivekeep.utils.collections
+
+inline fun <T> List<T>.ifNotEmpty(action: (collection: List<T>) -> Unit) {
+    if (this.isNotEmpty()) {
+        action(this)
+    }
+}